Tidying up some installation issues

+++ docs/conf.py
@@ -12,6 +12,8 @@ import time
 
 import git
 import sphinx_rtd_theme
+import stat
+
 from mpf.core.config_validator import ConfigValidator
 
 sys.path.insert(0, os.path.abspath(os.pardir))
@@ -27,10 +29,6 @@ mpf_examples_link_name = 'examples'
 
 
 def setup_mpf_examples_link():
-    try:
-        os.unlink(mpf_examples_link_name)
-    except FileNotFoundError:
-        pass
 
     if os.path.isdir(os.path.join(os.getcwd(), os.pardir, os.pardir, 'mpf-examples')):
         examples_root = os.path.join(os.getcwd(), os.pardir, os.pardir, 'mpf-examples')
@@ -43,8 +41,28 @@ def setup_mpf_examples_link():
 
     verify_version(os.path.join(examples_root, '_version.py'))
 
-    print("Creating '{}' link to {}".format(mpf_examples_link_name, examples_root))
-    os.symlink(examples_root, mpf_examples_link_name)
+    # See if symlink already exists, and goes to the right place.
+    exists = True
+    bad = True
+    try:    
+        linkinfo = os.stat(mpf_examples_link_name, follow_symlinks=False)
+        if (stat.S_ISLNK(linkinfo.st_mode)):
+            linkPath = os.readlink(mpf_examples_link_name)
+            # This sometimes has trouble on Windows because realpath doesn't canonise case of the drive letter
+            if os.path.realpath(linkPath) == os.path.realpath(examples_root):
+                bad = False
+    except FileNotFoundError:
+        exists = False
+
+    # If it exists, but isn't a link or is to the wrong place, delete it
+    if bad and exists:
+        os.unlink(mpf_examples_link_name)
+
+    # If it didn't exist or we just deleted it, created it
+    if bad or not exists:
+        # Flush to ensure message is displayed before potential system error from symlink failing
+        print("Creating '{}' link to {}".format(mpf_examples_link_name, examples_root), flush=True)
+        os.symlink(examples_root, mpf_examples_link_name)
 
 def verify_version(version_file):
